refactor(horizon-householder-appeal-publish): seperated concerns for handler to test FaaS
