+++ packages/horizon-householder-appeal-publish/handler.js
@@ -1,283 +1,10 @@
 const axios = require('axios');
-
-const config = {
-  appealsService: {
-    url: process.env.APPEALS_SERVICE_URL,
-  },
-  documentService: {
-    url: process.env.DOCUMENT_SERVICE_URL,
-  },
-  horizon: {
-    url: process.env.HORIZON_URL,
-  },
-  openfaas: {
-    gatewayUrl: process.env.GATEWAY_URL,
-  },
-};
-
-function convertToSoapKVPair(log, key, value) {
-  log.debug(
-    {
-      key,
-      value,
-    },
-    'Parsing soap key/value pair'
-  );
-
-  if (Array.isArray(value)) {
-    log.debug('Parsing as array');
-    return {
-      'a:AttributeValue': {
-        '__i:type': 'a:SetAttributeValue',
-        'a:Name': key,
-        'a:Values': value.map((item) => convertToSoapKVPair(log, item.key, item.value)),
-      },
-    };
-  }
-
-  if (value == null) {
-    log.debug('Parsing as null');
-    return {
-      'a:AttributeValue': {
-        '__i:type': 'a:StringAttributeValue',
-        'a:Name': key,
-        'a:Value': null,
-      },
-    };
-  }
-
-  if (value.toISOString) {
-    /* Value is a date */
-    log.debug('Parsing as date');
-    return {
-      'a:AttributeValue': {
-        '__i:type': 'a:DateAttributeValue',
-        'a:Name': key,
-        'a:Value': value.toISOString(),
-      },
-    };
-  }
-
-  let cleanValue = value;
-
-  /* Check for booleans - convert to "Yes" or "No" */
-  if (cleanValue === true) {
-    cleanValue = 'Yes';
-  } else if (cleanValue === false) {
-    cleanValue = 'No';
-  }
-
-  /* Everything else is a string */
-  log.debug('Parsing as string');
-  return {
-    'a:AttributeValue': {
-      '__i:type': 'a:StringAttributeValue',
-      'a:Name': key,
-      'a:Value': `${cleanValue}`, // Ensure value a string
-    },
-  };
-}
-
-/**
- * Call Horizon
- *
- * Calls Horizon and gets the ID of the created
- * case
- *
- * @param log
- * @param input
- * @returns {Promise<string>}
- */
-async function callHorizon(log, input) {
-  log.info(input, 'Calling Horizon');
-
-  const { data } = await axios.post('/horizon', input, {
-    baseURL: config.horizon.url,
-  });
-
-  log.info({ data }, 'Horizon response');
-
-  // case IDs are in format APP/W4705/D/21/3218521 - we need last 7 digits or numbers after final slash (always the same)
-  const horizonFullCaseId = data?.Envelope?.Body?.CreateCaseResponse?.CreateCaseResult?.value;
-
-  if (!horizonFullCaseId) {
-    log.error(
-      { input: data?.Envelope?.Body?.CreateCaseResponse?.CreateCaseResult },
-      'Horizon ID malformed'
-    );
-    throw new Error('Horizon ID malformed');
-  }
-
-  const parsedHorizonId = horizonFullCaseId.split('/').slice(-1).pop();
-
-  log.debug({ parsedHorizonId, horizonFullCaseId }, 'Horizon ID parsed');
-
-  return parsedHorizonId;
-}
-
-/**
- * Create Contacts
- *
- * Takes the body data, parses it into agent/appellant, creates the
- * contact inside Horizon and returns the data that is in the format
- * that can be used by the CreateCase endpoint.
- *
- * @param {*} log
- * @param {*} body
- * @returns {Promise}
- */
-async function createContacts(log, body) {
-  const contacts = [];
-  let logMessage;
-
-  if (body.appeal.aboutYouSection.yourDetails.isOriginalApplicant) {
-    /* User is original applicant - just add appellant */
-    logMessage = 'User is original applicant';
-    contacts.push({
-      type: 'Appellant',
-      email: body.appeal.aboutYouSection.yourDetails.email,
-      name: body.appeal.aboutYouSection.yourDetails.name,
-    });
-  } else {
-    /* User is agent - add both agent and OP */
-    logMessage = 'User is agent';
-    contacts.push(
-      {
-        type: 'Agent',
-        email: body.appeal.aboutYouSection.yourDetails.email,
-        name: body.appeal.aboutYouSection.yourDetails.name,
-      },
-      {
-        /* Email not collected here */
-        type: 'Appellant',
-        name: body.appeal.aboutYouSection.yourDetails.appealingOnBehalfOf,
-      }
-    );
-  }
-
-  log.info(contacts, logMessage);
-
-  return Promise.all(
-    contacts.map(async ({ name, email, type }) => {
-      /* Create the user in Horizon */
-      const [firstName, ...lastName] = name.split(' ');
-
-      log.info('Inserting contact into Horizon');
-
-      const {
-        data: { id: contactId },
-      } = await axios.post(
-        '/function/horizon-create-contact',
-        {
-          firstName,
-          lastName: lastName.join(' '), // Treat multiple spaces as part of last name
-          email,
-        },
-        {
-          baseURL: config.openfaas.gatewayUrl,
-        }
-      );
-
-      return {
-        /* Add user contact details */
-        key: 'Case Involvement:Case Involvement',
-        value: [
-          {
-            key: 'Case Involvement:Case Involvement:ContactID',
-            value: contactId,
-          },
-          {
-            key: 'Case Involvement:Case Involvement:Contact Details',
-            value: name,
-          },
-          {
-            key: 'Case Involvement:Case Involvement:Involvement Start Date',
-            value: new Date(),
-          },
-          {
-            key: 'Case Involvement:Case Involvement:Communication Preference',
-            value: 'e-mail',
-          },
-          {
-            key: 'Case Involvement:Case Involvement:Type Of Involvement',
-            value: type,
-          },
-        ],
-      };
-    })
-  );
-}
-
-/**
- * Get LPA Data
- *
- * Returns the LPA data
- *
- * @param log
- * @param code
- * @returns {Promise<{ id: string, name: string, inTrial: boolean, england: boolean, wales: boolean, horizonId: string }>}
- */
-async function getLpaData(log, code) {
-  log.info({ code }, 'Getting LPA data from Appeals Service API');
-
-  const { data } = await axios.get(`/api/v1/local-planning-authorities/${code}`, {
-    baseURL: config.appealsService.url,
-  });
-
-  if (!data?.horizonId) {
-    log.error({ data }, 'LPA not found');
-    throw new Error('Unknown LPA');
-  }
-
-  return data;
-}
-
-/**
- * Publish Documents
- *
- * Publishes the documents to Horizon. This is done asynchronously
- * as we're not interested in the response
- *
- * @param log
- * @param {{id: string, type: string}[]} documents
- * @param {string} appealId
- * @param {string} horizonId
- * @returns {Promise<void>}
- */
-async function publishDocuments(log, documents, appealId, horizonId) {
-  await Promise.all(
-    documents
-      /* Remove any undefined keys */
-      .filter(({ id }) => id)
-      .map(async ({ id: documentId, type: documentType }) => {
-        log.info(
-          {
-            documentId,
-            documentType,
-            horizonId,
-            appealId,
-          },
-          'Publish document to Horizon'
-        );
-
-        await axios.post(
-          '/async-function/horizon-add-document',
-          {
-            documentId,
-            documentType,
-            // These are named as-per Horizon keys
-            caseReference: horizonId,
-            applicationId: appealId,
-          },
-          {
-            baseURL: config.openfaas.gatewayUrl,
-          }
-        );
-
-        log.debug('Publish document request accepted');
-      })
-  );
-}
+const config = require('./src/config');
+const { convertToSoapKVPair } = require('./src/convertToSoapKVPair');
+const { callHorizon } = require('./src/callHorizon');
+const { createContacts } = require('./src/createContacts');
+const { getLpaData } = require('./src/getLpaData');
+const { publishDocuments } = require('./src/publishDocuments');
 
 module.exports = async (event, context) => {
   event.log.info({ config }, 'Received householder appeal publish request');
