+++ packages/horizon-householder-appeal-publish/src/config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  appealsService: {
+    url: process.env.APPEALS_SERVICE_URL,
+  },
+  documentService: {
+    url: process.env.DOCUMENT_SERVICE_URL,
+  },
+  horizon: {
+    url: process.env.HORIZON_URL,
+  },
+  openfaas: {
+    gatewayUrl: process.env.GATEWAY_URL,
+  },
+};
